don't always remove docker images in factory reset
If we are just restarting setup from scratch it wastes a lot of time
to re-download all the docker images.

+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/factory-reset-DANGER.sh
@@ -6,7 +6,7 @@ echo "FACTORY RESET" >> /run/adsb-feeder-image.log
 /opt/adsb/docker-compose-adsb down
 rm -f /opt/adsb/config/{.env,config.json};
 rm -f /opt/adsb/init-complete
-docker system prune -a -f
+[ "$1" = "-prune" ] && docker system prune -a -f
 
 echo "FACTORY RESET DONE" >> /run/adsb-feeder-image.log
